revert #1962 because of failure to correctly print errors in non compiling definitions
