+++ k-distribution/src/test/java/org/kframework/debugger/TstKDebugOnKORE_IT.java
@@ -35,12 +35,12 @@ public class TstKDebugOnKORE_IT {
     @Before
     public void setup() throws URISyntaxException, IOException {
         fileName = "/convertor-tests/kore_imp.k";
-        utils = new KoreUtils(fileName, "IMP", "IMP-SYNTAX", false);
+        utils = new KoreUtils(fileName, "IMP", "IMP-SYNTAX");
         trans = new KtoKORE();
         kRunOptions = new KRunOptions();
         String pgm = "int s, n; n = 10; while(0<=n) { s = s + n; n = n + -1; }";
         parsed = utils.getParsed(pgm, Source.apply("generated by " + getClass().getSimpleName()));
-        debuggerSession = new KoreKDebug(parsed, new KoreUtils(fileName, "IMP", "IMP-SYNTAX", false).getRewriter(), new Integer(500), null, utils.kem, kRunOptions, utils.compiledDef);
+        debuggerSession = new KoreKDebug(parsed, new KoreUtils(fileName, "IMP", "IMP-SYNTAX").getRewriter(), new Integer(500), null, utils.kem, kRunOptions, utils.compiledDef);
     }
 
     @Test @Ignore
