+++ k-distribution/src/test/java/org/kframework/kore/compile/ConfigurationInheritanceTest.java
@@ -21,7 +21,7 @@ public class ConfigurationInheritanceTest {
     @Test
     public void simpleTest() throws URISyntaxException, IOException {
         String filename = "/compiler-tests/configuration-inheritance.k";
-        KoreUtils utils = new KoreUtils(filename, "C", "C", false);
+        KoreUtils utils = new KoreUtils(filename, "C", "C");
         System.out.println(utils.compiledDef.executionModule().sentences().mkString("\n"));
 
         K kResult = utils.stepRewrite(utils.getParsed("x", Source.apply("generated by " + getClass().getSimpleName())), Optional.<Integer>empty());
