+++ k-distribution/src/test/java/org/kframework/kore/convertors/TstBackendOnKORE_IT.java
@@ -29,9 +29,9 @@ public class TstBackendOnKORE_IT {
     }
 
     @Test
-    public void kore_imp() throws URISyntaxException {
+    public void kore_imp() throws IOException, URISyntaxException {
         String filename = "/convertor-tests/" + name.getMethodName() + ".k";
-        KoreUtils utils = new KoreUtils(filename, "IMP", "IMP-SYNTAX", false);
+        KoreUtils utils = new KoreUtils(filename, "IMP", "IMP-SYNTAX");
 
         String pgm = "int s, n; n = 10; while(0<=n) { s = s + n; n = n + -1; }";
 
@@ -46,9 +46,9 @@ public class TstBackendOnKORE_IT {
     }
 
     @Test
-    public void kore_varlabel() throws URISyntaxException {
+    public void kore_varlabel() throws IOException, URISyntaxException {
         String filename = "/convertor-tests/" + name.getMethodName() + ".k";
-        KoreUtils utils = new KoreUtils(filename, "VARLABEL", "VARLABEL-SYNTAX", false);
+        KoreUtils utils = new KoreUtils(filename, "VARLABEL", "VARLABEL-SYNTAX");
 
         String pgm = "holder(1 +1)";
