+++ k-distribution/src/test/java/org/kframework/kore/convertors/TstTinyOnKORE_IT.java
@@ -72,7 +72,7 @@ public class TstTinyOnKORE_IT {
                 "<top><k>foo</k></top>");
     }
 
-    private void executeTest(String mainModule, String mainSyntaxModule, String programText, String expected) throws URISyntaxException, IOException {
+    private void executeTest(String mainModule, String mainSyntaxModule, String programText, String expected) throws URISyntaxException {
         String filename = "/convertor-tests/" + name.getMethodName() + ".k";
 
         File definitionFile = testResource(filename);
