+++ k-distribution/tests/regression/error-messages/x.k
@@ -1,3 +1,14 @@
 // Copyright (c) 2015 K Team. All Rights Reserved.
 module X-SYNTAX endmodule
 module X endmodule
+// circular-module-imports/a.k.out
+// context-validity/b.k.out
+// context-validity/c.k.out
+// context-validity/d.k.out
+// incorrect-stream-name/b.k.out
+// main-module-not-found/a.k.out
+// module-not-found/a.k.out
+// non-strict-sorts/b.k.out
+// non-strict-sorts/c.k.out
+// side-conditions-are-not-allowed-in-macro-rules/b.k.out
+// unsupported-io-unblock-patterns/b.k.out
