+++ kernel/src/main/java/org/kframework/Definition.java
@@ -51,24 +51,24 @@ public class Definition {
      * Parses the text to create a {@link Definition} object.
      */
     public static org.kframework.definition.Definition from(String definitionText, String mainModuleName, Source source, List<File> lookupDirectories) {
-        File tempDir = Files.createTempDir();
-        File theFileUtilTempDir = new File(tempDir.getAbsolutePath() + File.pathSeparator + "tempDir");
-        File definitionDir = new File(tempDir.getAbsolutePath() + File.pathSeparator + "definitionDir");
-        File workingDir = new File(tempDir.getAbsolutePath() + File.pathSeparator + "workingDir");
-        File kompiledDir = new File(tempDir.getAbsolutePath() + File.pathSeparator + "kompiledDir");
-        if(!theFileUtilTempDir.mkdir() || !definitionDir.mkdir() || !workingDir.mkdir() || !kompiledDir.mkdir()) {
+        File tmpDir = Files.createTempDir();
+        File tempDir = new File(tmpDir.getAbsolutePath() + File.pathSeparator + "tempDir");
+        File definitionDir = new File(tmpDir.getAbsolutePath() + File.pathSeparator + "definitionDir");
+        File workingDir = new File(tmpDir.getAbsolutePath() + File.pathSeparator + "workingDir");
+        File kompiledDir = new File(tmpDir.getAbsolutePath() + File.pathSeparator + "kompiledDir");
+        if(!tempDir.mkdir() || !definitionDir.mkdir() || !workingDir.mkdir() || !kompiledDir.mkdir()) {
             throw new AssertionError("Could not create one of the temporary directories");
         }
         GlobalOptions globalOptions = new GlobalOptions();
         KExceptionManager kem = new KExceptionManager(globalOptions);
 
-        FileUtil fileUtil = new FileUtil(theFileUtilTempDir,
+        FileUtil fileUtil = new FileUtil(tempDir,
                 Providers.of(definitionDir),
                 workingDir,
                 Providers.of(kompiledDir),
                 globalOptions,
                 System.getenv());
-        ParserUtils parserUtils = new ParserUtils(fileUtil::resolveWorkingDirectory, kem, globalOptions);
+        ParserUtils parserUtils = new ParserUtils(fileUtil, kem, globalOptions);
 
         org.kframework.definition.Definition definition = parserUtils.loadDefinition(
                 mainModuleName,
