+++ kernel/src/main/java/org/kframework/Parser.java
@@ -2,18 +2,16 @@
 package org.kframework;
 
 import org.kframework.attributes.Source;
-import org.kframework.definition.*;
+import org.kframework.definition.Module;
 import org.kframework.kore.K;
 import org.kframework.kore.Sort;
 import org.kframework.parser.concrete2kore.ParseInModule;
-import org.kframework.parser.concrete2kore.generator.RuleGrammarGenerator;
 import org.kframework.utils.errorsystem.ParseFailedException;
 import scala.Option;
 import scala.Tuple2;
 import scala.util.Either;
 import scala.util.control.Exception;
 
-import java.io.IOException;
 import java.util.HashSet;
 import java.util.Set;
 
@@ -26,20 +24,14 @@ public class Parser {
     private final ParseInModule parseInModule;
 
     private Parser(Module module) {
-        // TODO: remove hack once the frontend is cleaner; also remove the IOException once the hack is cleared
-        if (module.name().endsWith(RuleGrammarGenerator.RULE_CELLS)) {
-            org.kframework.definition.Definition definitionWithBuiltins = Definition.from("require \"domians.k\"", "K");
-            this.parseInModule = new RuleGrammarGenerator(definitionWithBuiltins, true).getCombinedGrammar(module);
-        } else {
-            this.parseInModule = new ParseInModule(module);
-        }
+        this.parseInModule = new ParseInModule(module);
     }
 
     /**
      * Parses a string with a particular start symbol/sort.
      *
      * @param startSymbol the start symbol/sort
-     * @param toParse     the String to parse
+     * @param toParse  the String to parse
      * @param fromSource  the Source of the String toParse
      * @return a pair: the left projection is a parsed string as a K, if successful;
      * the right projection is the set of issues encountered while parsing
@@ -59,7 +51,7 @@ public class Parser {
      * Parses a string with a particular start symbol/sort.
      *
      * @param startSymbol the start symbol/sort
-     * @param toParse     the String to parse
+     * @param toParse  the String to parse
      * @return a pair: the left projection is a parsed string as a K, if successful;
      * the right projection is the set of issues encountered while parsing
      */
