+++ kernel/src/test/java/org/kframework/kore/compile/GenerateSentencesFromConfigDeclTest.java
@@ -42,7 +42,7 @@ public class GenerateSentencesFromConfigDeclTest {
     public void setUp() {
         String definitionText;
         FileUtil files = FileUtil.testFileUtil();
-        ParserUtils parser = new ParserUtils(files::resolveWorkingDirectory, new KExceptionManager(new GlobalOptions()));
+        ParserUtils parser = new ParserUtils(files, new KExceptionManager(new GlobalOptions()));
         File definitionFile = new File(Kompile.BUILTIN_DIRECTORY.toString() + "/kast.k");
         definitionText = files.loadFromWorkingDirectory(definitionFile.getPath());
