+++ kernel/src/test/java/org/kframework/parser/concrete2kore/RuleGrammarTest.java
@@ -12,7 +12,6 @@ import org.kframework.attributes.Source;
 import org.kframework.definition.Definition;
 import org.kframework.definition.Module;
 import org.kframework.definition.RegexTerminal;
-import org.kframework.kompile.DefinitionParsing;
 import org.kframework.kompile.Kompile;
 import org.kframework.kore.K;
 import org.kframework.kore.Sort;
@@ -32,7 +31,7 @@ import java.util.Set;
 import static org.kframework.kore.KORE.*;
 
 public class RuleGrammarTest {
-    private final static Sort startSymbol = DefinitionParsing.START_SYMBOL;
+    private final static Sort startSymbol = Kompile.START_SYMBOL;
     private RuleGrammarGenerator gen;
 
     @Before
@@ -43,7 +42,7 @@ public class RuleGrammarTest {
     public RuleGrammarGenerator makeRuleGrammarGenerator() {
         String definitionText;
         FileUtil files = FileUtil.testFileUtil();
-        ParserUtils parser = new ParserUtils(files::resolveWorkingDirectory, new KExceptionManager(new GlobalOptions()));
+        ParserUtils parser = new ParserUtils(files, new KExceptionManager(new GlobalOptions()));
         File definitionFile = new File(Kompile.BUILTIN_DIRECTORY.toString() + "/kast.k");
         definitionText = files.loadFromWorkingDirectory(definitionFile.getPath());
